fixed Goals + changed all user id into long

+++ SocialApp/SocialApp/Services/GroceryListService.cs
@@ -19,7 +19,7 @@
             this.gR = gR;
         }
 
-        public async Task<List<GroceryIngredient>> GetIngredientsForUser(int userId)
+        public async Task<List<GroceryIngredient>> GetIngredientsForUser(long userId)
         {
             var ingredients = await this.groceryRepository.GetIngredientsForUser(userId);
             foreach (var ingredient in ingredients)
@@ -37,12 +37,12 @@
             return ingredients;
         }
 
-        public async Task UpdateIsChecked(int userId, int ingredientId, bool isChecked)
+        public async Task UpdateIsChecked(long userId, int ingredientId, bool isChecked)
         {
             await this.groceryRepository.UpdateIsChecked(userId, ingredientId, isChecked);
         }
 
-        public async Task<GroceryIngredient> AddIngredientToUser(int userId, GroceryIngredient ingredient, string newGroceryIngredientName, ObservableCollection<SectionModel> sections)
+        public async Task<GroceryIngredient> AddIngredientToUser(long userId, GroceryIngredient ingredient, string newGroceryIngredientName, ObservableCollection<SectionModel> sections)
         {
             if (ingredient == GroceryIngredient.defaultIngredient)
             {
